bugs fixed

+++ frontend/src/app/pages/admin/orders/orders.component.ts
@@ -1,5 +1,4 @@
 import { Component } from '@angular/core';
-import { DeleteAlertService } from '../../../shared/components/delete-alert/delete-alert.service';
 import { Order } from '../../../shared/interfaces/order.interface';
 import { OrdersService } from './orders.service';
 
@@ -9,23 +8,16 @@ import { OrdersService } from './orders.service';
   styleUrls: ['./orders.component.scss'],
 })
 export class OrdersComponent {
-  constructor(
-    private deleteAlertService: DeleteAlertService,
-    private ordersService: OrdersService
-  ) {
+  constructor(private ordersService: OrdersService) {
     this.getOrders();
   }
 
   loading = true;
-
   showOrder = false;
   orderToShow: Order;
-
   orders: Order[];
-
-  mobileOrders = [];
-
   deleteAlert: boolean;
+  deleteIndex: number;
 
   getOrders() {
     this.orders = [];
@@ -37,7 +29,8 @@ export class OrdersComponent {
   }
 
   openOrder(order: Order) {
-    if (!this.deleteAlert) {  //! you can do it better, don't emit openOrder!
+    if (!this.deleteAlert) {
+      //! you can do it better, don't emit openOrder!
       this.orderToShow = order;
       this.showOrder = true;
     }
@@ -56,21 +49,19 @@ export class OrdersComponent {
     this.ordersService.updateOrder(order._id, status).subscribe();
   }
 
-  onDelete(index: number) {
+  onDelete(confirmed) {
+    const order = this.orders[this.deleteIndex]._id;
+    if (confirmed) {
+      this.ordersService.deleteOrder(order).subscribe(() => {
+        this.orders.splice(this.deleteIndex, 1);
+      });
+      this.deleteAlert = false;
+    } else {
+      this.deleteAlert = false;
+    }
+  }
+  openDeleteAlert(index) {
+    this.deleteIndex = index;
     this.deleteAlert = true;
-    const order = this.orders[index]._id;
-    this.deleteAlertService.deleteMessage.subscribe((response) => {
-      switch (response) {
-        case true:
-          this.ordersService.deleteOrder(order).subscribe(() => {
-            this.orders.splice(index, 1);
-          });
-          this.deleteAlert = false;
-          break;
-        case false:
-          this.deleteAlert = false;
-          break;
-      }
-    });
   }
 }
